chore: refactor talpaCSV columns into own class
